% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R, R/old_analyze.R
\name{analyze}
\alias{analyze}
\title{Correction of Genomes in R}
\usage{
analyze(i = double(), j = double())

analyze(i = double(), j = double())
}
\arguments{
\item{i}{i index.}

\item{j}{j index.}

\item{path.base}{Base of the path.}

\item{summary.file}{Path to summary file.}

\item{output}{Output stream to write to.}

\item{T}{testing (only read one gen file)}

\item{safe}{Don't delete files}

\item{i}{Index 1}

\item{j}{Index 2}

\item{k}{Index 3}
}
\value{
Flat file at specified path.

Output file at container.
}
\description{
Software for the Examination of Multiple Correction Methodologies in Accurate Genomic Environments

Software for the Examination of Multiple Correction Methodologies in Accurate Genomic Environments
}

