% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_analyze.R
\name{analyze_old}
\alias{analyze_old}
\title{Correction of Genomes in R}
\usage{
analyze_old(i = double(), j = double())
}
\arguments{
\item{i}{Index 1}

\item{j}{Index 2}

\item{k}{Index 3}
}
\value{
Output file at container.
}
\description{
Software for the Examination of Multiple Correction Methodologies in Accurate Genomic Environments
}

