% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_phen_df.R
\name{assoc_wrapper}
\alias{assoc_wrapper}
\title{Gen Phen to Df in preparation for Lm}
\usage{
assoc_wrapper(gen, samp)
}
\description{
Takes as input genotype and phenotype files and outputs a df ready to be analyzed through lm
}

