% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2gen2.R
\name{calculate_was}
\alias{calculate_was}
\title{Takes gen file and calculates weighted allele scores (WAS)}
\usage{
calculate_was(gen = NULL, snps = NULL)
}
\arguments{
\item{gen}{Truncated gen file with only causal snps}

\item{snps}{Truncated summary file with only causal snps}
}
\value{
A vector of weighted allele scores
}
\description{
Takes gen file and calculates weighted allele scores (WAS)
}

