% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.snps.R
\name{causal.snps}
\alias{causal.snps}
\title{Function for obtaining a vector of causal SNPs.}
\usage{
causal.snps(.summary = NULL, mode = "default", nc = NULL, maf = FALSE,
  maf_range = NULL)
}
\arguments{
\item{.summary}{List of SNPs with attributes}

\item{mode}{Mode}

\item{nc}{Number of causal SNPs}

\item{maf}{TRUE is selecting via maf}

\item{maf_range}{Range of the form c(lower, upper)}
}
\description{
Function for obtaining a vector of causal SNPs.
}

