% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake.snps.R
\name{fake.snps}
\alias{fake.snps}
\title{Create a vector of fake non-sense SNPs}
\usage{
fake.snps(summary = NULL, n = 3000, group = NULL)
}
\arguments{
\item{summary}{Summary sheet with info on SNPs}

\item{n}{Number to generate}

\item{group}{Optional grouping paramter. Reports as a list}
}
\description{
Create a vector of fake non-sense SNPs
}

