% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr}
\alias{fdr}
\title{Calculates FDR given a data frame}
\usage{
fdr(df = NULL)
}
\arguments{
\item{df}{Data frame of input}
}
\value{
A vector of results
}
\description{
Calculates FDR given a data frame
}

