% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_gen2r.R
\name{gen2r}
\alias{gen2r}
\title{Oxford style genotype conversion}
\usage{
gen2r(genfile, local = TRUE)
}
\arguments{
\item{genfile}{File to convert}

\item{local}{TRUE is the file is already in the R workspace. FALSE if the file is a path which must be read in.}
}
\description{
Oxford style genotype conversion
}

