% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_phen_df.R
\name{gen_phen_df}
\alias{gen_phen_df}
\title{Gen Phen to Df in preparation for Lm}
\usage{
gen_phen_df(gen, samp)
}
\description{
Takes as input genotype and phenotype files and outputs a df ready to be analyzed through lm
}

