% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen.R
\name{phen}
\alias{phen}
\title{Phenotype Calculation}
\usage{
phen(.snps = snps, .combR = combR, .h2 = h2)
}
\arguments{
\item{.snps}{SNPs matrix}

\item{.combR}{combR matrix}
}
\description{
Phenotype Calculation
}

