% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Random Number Generator}
\usage{
rand(n = NULL, sum = NULL, start = 0)
}
\arguments{
\item{n}{Number of values to simulate}

\item{sum}{Sum of n numbers}

\item{start}{Lower bound of number generation.}
}
\value{
A vector of n numbers which add up to sum
}
\description{
Generates n numbers which add up to sum
}

