% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand0.R
\name{rand0}
\alias{rand0}
\title{Delta Generator Function}
\usage{
rand0()
}
\value{
Delta which either 1 or -1 with equal probability.
}
\description{
Creates a random number which is either -1 or 1 with .5 probability of either.
}

