% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.gen.R
\name{sim.gen}
\alias{sim.gen}
\title{Haplotpe Simulation Function}
\usage{
sim.gen(i = integer(), j = integer(), nj = integer(),
  container = "/scratch/hpc2862/CAMH/perm_container/")
}
\arguments{
\item{i}{Index 1}

\item{j}{Index 2}

\item{nj}{Number of jobs}

\item{container}{Where should the simulated genotypes be put? This should be a root diretory where the function can create nj x i containers to store the different runs.}
}
\value{
nj x i different folders stemming from the root directory specified by container
}
\description{
Generate random simulated haplotypes with Hapgen2 through qsub.
}

