% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Create strata}
\usage{
stratify(snp_list = NULL, summary = NULL, p = NULL, n_strata = NULL)
}
\arguments{
\item{snp_list}{Vector of true SNPs}

\item{summary}{Full summmary sheet}

\item{p}{Proportion in strata 1}

\item{n_strata}{Number of strata. Currently only supports 2.}
}
\description{
Currently only set up to handle two stata for arbitrary p in [0,1].
}

