% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{sub_job}
\alias{sub_job}
\title{Submission Convenience function}
\usage{
sub_job(n = NULL)
}
\arguments{
\item{n}{Number of jobs to submit. Defaults}
}
\value{
Nothing. Submits directly to QSUB; check with qstat.
}
\description{
Submission Convenience function
}

