% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R, R/analyze.new.R
\name{analyze}
\alias{analyze}
\title{Correction of Genomes in R}
\usage{
analyze(i = double(), j = double(),
  path.base = "/scratch/hpc2862/CAMH/perm_container/container_",
  summary.file = "/scratch/hpc2862/CAMH/perm_container/snp_summary2.out",
  output = "~/repos/coR-ge/data/test_run.txt")

analyze(i = double(), j = double(),
  path.base = "/scratch/hpc2862/CAMH/perm_container/container_",
  summary.file = "/scratch/hpc2862/CAMH/perm_container/snp_summary2.out",
  output = "~/repos/coR-ge/data/test_run.txt")
}
\arguments{
\item{i}{Index 1}

\item{j}{Index 2}

\item{path.base}{Base of the path.}

\item{summary.file}{Path to summary file.}

\item{output}{Output stream to write to.}

\item{k}{Index 3}

\item{i}{i index.}

\item{j}{j index.}
}
\value{
Output file at container.

SQL database at specified path.
}
\description{
Software for the Examination of Multiple Correction Methodologies in Accurate Genomic Environments

Software for the Examination of Multiple Correction Methodologies in Accurate Genomic Environments
}

