% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Correction of Genomes in R}
\usage{
analyze(i = double(), j = double(), mode = "default",
  path.base = "/scratch/hpc2862/CAMH/perm_container/container_",
  summary.file = "/scratch/hpc2862/CAMH/perm_container/snp_summary2.out",
  output = "~/repos/coR-ge/data/test_run2.txt", test = TRUE, safe = TRUE,
  local = FALSE, h2 = 0.45, pc = 0.5, pnc = 0.5)
}
\arguments{
\item{i}{i index.}

\item{j}{j index.}

\item{path.base}{Base of the path.}

\item{summary.file}{Path to summary file.}

\item{output}{Output stream to write to.}

\item{test}{testing (only read one gen file)}

\item{safe}{Don't delete files}

\item{pc}{Proportion of causal SNPs in the second strata}

\item{pnc}{Proportion of noncausal SNPs in the second strata}
}
\value{
Flat file at specified path.
}
\description{
Software for the Examination of Multiple Correction Methodologies in Accurate Genomic Environments
}

