% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.snps.R
\name{causal.snps}
\alias{causal.snps}
\title{Function for obtaining a vector of causal SNPs.}
\usage{
causal.snps(summary = NULL)
}
\arguments{
\item{summary}{List of SNPs with attributes}
}
\description{
Function for obtaining a vector of causal SNPs.
}

