% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\title{Correct strata}
\usage{
correct(strata = NULL, n_strata = NULL, assoc = NULL)
}
\arguments{
\item{strata}{Strata}

\item{n_strata}{Number of strata}

\item{assoc}{Assoc file path}
}
\description{
Correct strata
}

