% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_mutate.R
\name{SE_mutate}
\alias{SE_mutate}
\title{Automated standard evaluation of mutate with %in% operator.}
\usage{
SE_mutate(col1, col2, new_col_name, df)
}
\arguments{
\item{col1}{col1 %in% col2}

\item{col2}{col1 %in% col2}

\item{new_col_name}{Name of new column. Can be passed as variable.}

\item{df}{Data frame to append to.}
}
\value{
Data frame with ncol(df)+1 columns.
}
\description{
Automated standard evaluation of mutate with %in% operator.
}

